feat(ts): convert au-region-selector to TS
